refactor: move calendar to utils
