+++ utils/index.ts
@@ -2,5 +2,6 @@ import { logger } from './logs'
 import * as sql from './sql'
 import * as format from './format'
 import * as object from './object'
+import * as calendar from './calendar'
 
-export { logger, sql, format, object }
+export { logger, sql, format, object, calendar }
